+++ corehq/apps/api/resources/v0_4.py
@@ -28,8 +28,8 @@ from corehq.apps.app_manager.dbaccessors import get_apps_in_domain
 from corehq.apps.app_manager.models import Application, RemoteApp
 from corehq.apps.cloudcare.api import ElasticCaseQuery
 from corehq.apps.groups.models import Group
-from corehq.apps.repeaters.models import Repeater
-from corehq.apps.repeaters.utils import get_all_repeater_types
+from corehq.motech.repeaters.models import Repeater
+from corehq.motech.repeaters.utils import get_all_repeater_types
 from corehq.apps.users.models import CouchUser, Permissions
 from corehq.apps.users.util import format_username
 from couchforms.models import doc_types
