+++ corehq/apps/domain/tests/test_views.py
@@ -8,7 +8,7 @@ from django.test.client import Client
 from mock import patch
 
 from corehq.apps.users.models import WebUser
-from corehq.apps.repeaters.models import AppStructureRepeater
+from corehq.motech.repeaters.models import AppStructureRepeater
 from corehq.apps.domain.models import Domain
 from corehq.apps.app_manager.models import Application, APP_V1
 from corehq.apps.domain.views import CreateNewExchangeSnapshotView
