+++ corehq/motech/repeaters/forms.py
@@ -2,7 +2,7 @@ from django import forms
 from django.core.exceptions import ValidationError
 from django.utils.translation import ugettext_lazy as _
 
-from corehq.apps.repeaters.repeater_generators import RegisterGenerator
+from corehq.motech.repeaters.repeater_generators import RegisterGenerator
 from corehq.apps.reports.analytics.esaccessors import get_case_types_for_domain_es
 from crispy_forms.helper import FormHelper
 from crispy_forms import layout as crispy
