+++ corehq/motech/repeaters/tests/test_repeater.py
@@ -14,18 +14,18 @@ from corehq.apps.domain.shortcuts import create_domain
 from corehq.apps.locations.models import SQLLocation, LocationType
 from corehq.apps.receiverwrapper.exceptions import DuplicateFormatException, IgnoreDocument
 from corehq.apps.receiverwrapper.util import submit_form_locally
-from corehq.apps.repeaters.repeater_generators import FormRepeaterXMLPayloadGenerator, RegisterGenerator, \
+from corehq.motech.repeaters.repeater_generators import FormRepeaterXMLPayloadGenerator, RegisterGenerator, \
     BasePayloadGenerator
-from corehq.apps.repeaters.tasks import check_repeaters
-from corehq.apps.repeaters.models import (
+from corehq.motech.repeaters.tasks import check_repeaters
+from corehq.motech.repeaters.models import (
     CaseRepeater,
     FormRepeater,
     UserRepeater,
     LocationRepeater,
     RepeatRecord,
     ShortFormRepeater)
-from corehq.apps.repeaters.const import MIN_RETRY_WAIT, POST_TIMEOUT, RECORD_SUCCESS_STATE
-from corehq.apps.repeaters.dbaccessors import delete_all_repeat_records, delete_all_repeaters
+from corehq.motech.repeaters.const import MIN_RETRY_WAIT, POST_TIMEOUT, RECORD_SUCCESS_STATE
+from corehq.motech.repeaters.dbaccessors import delete_all_repeat_records, delete_all_repeaters
 from corehq.apps.users.models import CommCareUser
 from corehq.form_processor.tests.utils import run_with_all_backends, FormProcessorTestUtils
 from corehq.form_processor.interfaces.dbaccessors import CaseAccessors, FormAccessors
@@ -142,7 +142,7 @@ class RepeaterTest(BaseRepeaterTest):
 
         for repeat_record in repeat_records:
             with patch(
-                    'corehq.apps.repeaters.models.simple_post',
+                    'corehq.motech.repeaters.models.simple_post',
                     return_value=MockResponse(status_code=404, reason='Not Found')) as mock_post:
                 repeat_record.fire()
                 self.assertEqual(mock_post.call_count, 1)
@@ -179,7 +179,7 @@ class RepeaterTest(BaseRepeaterTest):
 
         for repeat_record in repeat_records:
             with patch(
-                    'corehq.apps.repeaters.models.simple_post',
+                    'corehq.motech.repeaters.models.simple_post',
                     return_value=MockResponse(status_code=200, reason='No Reason')) as mock_post:
                 repeat_record.fire()
                 self.assertEqual(mock_post.call_count, 1)
@@ -192,7 +192,7 @@ class RepeaterTest(BaseRepeaterTest):
                 )
 
         # The following is pretty fickle and depends on which of
-        #   - corehq.apps.repeaters.signals
+        #   - corehq.motech.repeaters.signals
         #   - casexml.apps.case.signals
         # gets loaded first.
         # This is deterministic but easily affected by minor code changes
@@ -213,11 +213,11 @@ class RepeaterTest(BaseRepeaterTest):
     def test_check_repeat_records(self):
         self.assertEqual(len(RepeatRecord.all()), 2)
 
-        with patch('corehq.apps.repeaters.models.simple_post') as mock_fire:
+        with patch('corehq.motech.repeaters.models.simple_post') as mock_fire:
             check_repeaters()
             self.assertEqual(mock_fire.call_count, 2)
 
-        with patch('corehq.apps.repeaters.models.simple_post') as mock_fire:
+        with patch('corehq.motech.repeaters.models.simple_post') as mock_fire:
             check_repeaters()
             self.assertEqual(mock_fire.call_count, 0)
 
@@ -229,13 +229,13 @@ class RepeaterTest(BaseRepeaterTest):
         for repeat_record in RepeatRecord.all():
             repeat_record.cancelled = True
             repeat_record.save()
-        with patch('corehq.apps.repeaters.models.simple_post') as mock_fire:
+        with patch('corehq.motech.repeaters.models.simple_post') as mock_fire:
             check_repeaters()
             self.assertEqual(mock_fire.call_count, 0)
 
         # trigger force send records if not cancelled and tries not exhausted
         for repeat_record in RepeatRecord.all():
-            with patch('corehq.apps.repeaters.models.simple_post',
+            with patch('corehq.motech.repeaters.models.simple_post',
                        return_value=MockResponse(status_code=200, reason='')
                        ) as mock_fire:
                 repeat_record.fire(force_send=True)
@@ -247,7 +247,7 @@ class RepeaterTest(BaseRepeaterTest):
                 self.assertEqual(repeat_record.overall_tries, 1)
 
         # not trigger records succeeded triggered after cancellation
-        with patch('corehq.apps.repeaters.models.simple_post') as mock_fire:
+        with patch('corehq.motech.repeaters.models.simple_post') as mock_fire:
             check_repeaters()
             self.assertEqual(mock_fire.call_count, 0)
             for repeat_record in RepeatRecord.all():
@@ -257,11 +257,11 @@ class RepeaterTest(BaseRepeaterTest):
     def test_process_repeat_record_locking(self):
         self.assertEqual(len(RepeatRecord.all()), 2)
 
-        with patch('corehq.apps.repeaters.tasks.process_repeat_record') as mock_process:
+        with patch('corehq.motech.repeaters.tasks.process_repeat_record') as mock_process:
             check_repeaters()
             self.assertEqual(mock_process.delay.call_count, 2)
 
-        with patch('corehq.apps.repeaters.tasks.process_repeat_record') as mock_process:
+        with patch('corehq.motech.repeaters.tasks.process_repeat_record') as mock_process:
             check_repeaters()
             self.assertEqual(mock_process.delay.call_count, 0)
 
@@ -270,7 +270,7 @@ class RepeaterTest(BaseRepeaterTest):
         for record in records:
             record.save()
 
-        with patch('corehq.apps.repeaters.tasks.process_repeat_record') as mock_process:
+        with patch('corehq.motech.repeaters.tasks.process_repeat_record') as mock_process:
             check_repeaters()
             self.assertEqual(mock_process.delay.call_count, 2)
 
@@ -278,10 +278,10 @@ class RepeaterTest(BaseRepeaterTest):
     def test_automatic_cancel_repeat_record(self):
         repeat_record = self.case_repeater.register(CaseAccessors(self.domain).get_case(CASE_ID))
         repeat_record.overall_tries = 1
-        with patch('corehq.apps.repeaters.models.simple_post', side_effect=Exception('Boom!')):
+        with patch('corehq.motech.repeaters.models.simple_post', side_effect=Exception('Boom!')):
             repeat_record.fire()
         self.assertEqual(2, repeat_record.overall_tries)
-        with patch('corehq.apps.repeaters.models.simple_post', side_effect=Exception('Boom!')):
+        with patch('corehq.motech.repeaters.models.simple_post', side_effect=Exception('Boom!')):
             repeat_record.fire()
         self.assertEqual(True, repeat_record.cancelled)
         repeat_record.requeue()
@@ -562,14 +562,14 @@ class RepeaterFailureTest(BaseRepeaterTest):
     @run_with_all_backends
     def test_failure(self):
         repeat_record = self.repeater.register(CaseAccessors(self.domain_name).get_case(CASE_ID))
-        with patch('corehq.apps.repeaters.models.simple_post', side_effect=Exception('Boom!')):
+        with patch('corehq.motech.repeaters.models.simple_post', side_effect=Exception('Boom!')):
             repeat_record.fire()
 
         self.assertEquals(repeat_record.failure_reason, 'Boom!')
         self.assertFalse(repeat_record.succeeded)
 
         # Should be marked as successful after a successful run
-        with patch('corehq.apps.repeaters.models.simple_post'):
+        with patch('corehq.motech.repeaters.models.simple_post'):
             repeat_record.fire()
 
         self.assertTrue(repeat_record.succeeded)
@@ -684,7 +684,7 @@ class TestRepeaterFormat(BaseRepeaterTest):
     @run_with_all_backends
     def test_new_format_payload(self):
         repeat_record = self.repeater.register(CaseAccessors(self.domain).get_case(CASE_ID))
-        with patch('corehq.apps.repeaters.models.simple_post') as mock_post:
+        with patch('corehq.motech.repeaters.models.simple_post') as mock_post:
             repeat_record.fire()
             headers = self.repeater.get_headers(repeat_record)
             mock_post.assert_called_with(
