+++ custom/enikshay/integrations/nikshay/views.py
@@ -1,4 +1,4 @@
-from corehq.apps.repeaters.views import AddCaseRepeaterView, AddCustomSOAPCaseRepeaterView
+from corehq.motech.repeaters.views import AddCaseRepeaterView, AddCustomSOAPCaseRepeaterView
 
 
 class RegisterNikshayPatientRepeaterView(AddCaseRepeaterView):
